Add SlideManager class for managing slides and enhance font handling in text editor
