+++ src/renderer/presenterView/renderer.js
@@ -1,5 +1,4 @@
 import "../js/sideBarResize";
-import "./js/slideSelector";
 import "./js/fileOpen";
 
 import "bootstrap";
